reorganize code for readability
