+++ addon/util-models/page.js
@@ -1,0 +1,6 @@
+import EmberObject from '@ember/object';
+
+export default EmberObject.extend({
+  number: 1
+});
+
